% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvregerrGS.R
\name{mvregerrGS}
\alias{mvregerrGS}
\title{Get Gibbs samples}
\usage{
mvregerrGS(Z, Z.errV, X, nchain = 100, init = list(nu = 5), every = 10,
  XP = NULL, record = c("Y"))
}
\arguments{
\item{Z}{\eqn{n\times k} matrix of observations}

\item{Z.errV}{\eqn{n\times k} matrix of observation errors (as variances)}

\item{X}{\eqn{n\times p} matrix of covariates (see details)}

\item{nchain}{chain length}

\item{init}{list of parameters determining prior (see details)}

\item{every}{how many iterations to report progress}

\item{XP}{covariates to use in making predictions}

\item{record}{character vector specifying which quantities to record and return}
}
\value{
a list of chains (see details)
}
\description{
This is the main function of the package. TODO
}
\details{
Details TODO
}
\author{
Pete Dodd
}
